verdant_parsers_git

Enum Lang

#[non_exhaustive]
pub enum Lang {

    Ada,
    Agda,
    Apex,
    Arduino,
    Asm,
    Asm68k,
    Astro,
    Authzed,
    Awk,
    Bash,
    Bass,
    Beancount,
    Bib,
    Bibtex,
    Bicep,
    Bitbake,
    Blade,
    Blueprint,
    Bp,
    Bpftrace,
    Brightscript,
    Bzl,
    C,
    C3,
    CSharp,
    Caddy,
    Cairo,
    Capnp,
    Chatito,
    Circom,
    Clarity,
    Clojure,
    Cmake,
    Comment,
    Cook,
    Corn,
    Cpon,
    Cpp,
    Csharp,
    Css,
    Csv,
    Cuda,
    Cue,
    Cylc,
    D,
    Dart,
    Desktop,
    Dhall,
    Diff,
    Disassembly,
    Djot,
    Dockerfile,
    Dot,
    Doxygen,
    Dtd,
    Dts,
    Earthfile,
    Ebnf,
    Editorconfig,
    Eds,
    Eelixir,
    Ejs,
    Elisp,
    Elixir,
    Elm,
    Elsa,
    Elvish,
    Enforce,
    Erb,
    Erlang,
    Eruby,
    Faust,
    Fennel,
    Fidl,
    Firrtl,
    Fish,
    Foam,
    Forth,
    Fortran,
    Fsd,
    Fsh,
    Fsharp,
    Func,
    Fusion,
    Gap,
    Gaptst,
    Gdresource,
    Gdscript,
    Gdshaderinc,
    Gitattributes,
    Gitcommit,
    Gitconfig,
    Gitdiff,
    Gitignore,
    Gitrebase,
    Gleam,
    GlimmerJavascript,
    GlimmerTypescript,
    Glsl,
    Gn,
    Gnuplot,
    Go,
    Goctl,
    Gomod,
    Gosum,
    Gotmpl,
    Gowork,
    Gpg,
    Graphql,
    Gren,
    Groovy,
    Groq,
    Gstlaunch,
    Hack,
    Handlebars,
    Hare,
    Haskell,
    Haskellpersistent,
    Haxe,
    Hcl,
    Heex,
    Helm,
    Help,
    Hexdump,
    Hjson,
    Hlsl,
    Hlsplaylist,
    Hocon,
    Hoon,
    Html,
    Htmlangular,
    Htmldjango,
    Http,
    Hurl,
    Hyprlang,
    Idl,
    Idris2,
    Ini,
    Inko,
    Ispc,
    Janet,
    Java,
    Javadoc,
    Javascript,
    Jinja,
    JinjaInline,
    Jproperties,
    Jq,
    Jsdoc,
    Json,
    Json5,
    Jsonc,
    Jsonnet,
    Julia,
    Just,
    Kcl,
    Kconfig,
    Kdl,
    Kitty,
    Kos,
    Kotlin,
    Koto,
    Kusto,
    Lalrpop,
    Latex,
    Ld,
    Ledger,
    Leo,
    Liquid,
    Liquidsoap,
    Lisp,
    Llvm,
    Lua,
    Luadoc,
    Luap,
    Luau,
    Magik,
    Make,
    Markdown,
    MarkdownInline,
    Matlab,
    Menhir,
    Mermaid,
    Meson,
    Mlir,
    Muttrc,
    Nasm,
    Nginx,
    Nickel,
    Nim,
    NimFormatString,
    Ninja,
    Nix,
    Nqc,
    Nu,
    Objc,
    Objdump,
    Ocaml,
    OcamlInterface,
    Ocamlinterface,
    Ocamllex,
    Odin,
    Pascal,
    Passwd,
    Pbtxt,
    Pem,
    Perl,
    Pgn,
    Php,
    PhpOnly,
    Phpdoc,
    Pioasm,
    Pkl,
    Po,
    Pod,
    Poefilter,
    Pony,
    Printf,
    Prisma,
    Problog,
    Prolog,
    Promql,
    Proto,
    Prql,
    Ps1,
    Psv,
    Pug,
    Puppet,
    Purescript,
    Pymanifest,
    Python,
    Ql,
    Qml,
    Qmldir,
    Query,
    R,
    Racket,
    Ralph,
    Rasi,
    Razor,
    Rbs,
    Re2c,
    Readline,
    Regex,
    Rego,
    Requirements,
    Rescript,
    Rifleconf,
    Rnoweb,
    Robot,
    Robots,
    Roc,
    Ron,
    Rst,
    Ruby,
    Runescript,
    Rush,
    Rust,
    Scala,
    Scfg,
    Scheme,
    Scss,
    Sface,
    Sflog,
    Shaderslang,
    Slim,
    Slint,
    Smali,
    Smithy,
    Snakemake,
    Solidity,
    Soql,
    Sosl,
    Sourcepawn,
    Sparql,
    Sproto,
    Sql,
    Squirrel,
    Sshconfig,
    Strace,
    Styled,
    Supercollider,
    Superhtml,
    Svelte,
    Sway,
    Swift,
    Sxhkdrc,
    Systemtap,
    Tablegen,
    Tact,
    Tal,
    Tape,
    Tcl,
    Teal,
    Templ,
    Tera,
    Terraform,
    Tex,
    Thrift,
    Tiger,
    Tla,
    Tmux,
    Todotxt,
    Toml,
    Trace32,
    Tsv,
    Tsx,
    Turtle,
    Twig,
    Typescript,
    Typespec,
    Typoscript,
    Typst,
    Udevrules,
    Ungrammar,
    Unison,
    Ursa,
    Usd,
    V,
    Vala,
    Verilog,
    Vhdl,
    Vim,
    Vrl,
    Vto,
    Vue,
    Wat,
    Wgsl,
    WgslBevy,
    Wing,
    Wit,
    Wxml,
    Xcompose,
    Xdefaults,
    Xml,
    Yaml,
    Yang,
    Yuck,
    Zathurarc,
    Zig,
    Ziggy,
    ZiggySchema,
    Zsh,

}

An enum of every supported language in the current feature set.

An instance of the respective tree-stter Language can be obtained with the get method.

You can also get a Lang from its name using for_name, or for a FileType using for_file_type. See the docs for each variant to see its “name” and the supported file types. Both of these require the SupportedLanguage trait to be in scope.

See LANGUAGES for a list containing all variants and LANGUAGE_NAMES for a list of all valid names.

The enum is marked as non-exhaustive for two reasons:

New languages may be added in the future

The variants are enabled/disabled by features

Example

use verdant_parsers_git::{Lang, LANGUAGES, LANGUAGE_NAMES};
use verdant_core::language_set::{SupportedLanguage, FileType};

// you can get a `Lang` from its name
assert_eq!(Lang::Rust, Lang::for_name("rust", &()).unwrap());
// and for a file type
assert_eq!(Some(Lang::Rust), Lang::for_file_type(FileType::Rust, &()));

// `LANGUAGES` is a list of all variants,
// `LANGUAGE_NAMES` is a list of all variant names
for (&lang, &name) in LANGUAGES.iter().zip(LANGUAGE_NAMES) {
    assert_eq!(lang, Lang::for_name(name, &()).unwrap());

    // `Lang` instances can be turned into strings
    assert_eq!(lang, Lang::for_name(<Lang as SupportedLanguage<'_, ()>>::name(&lang), &()).unwrap());
    assert_eq!(lang, Lang::for_name(lang.to_string(), &()).unwrap());
    assert_eq!(lang, Lang::for_name(lang.as_ref(), &()).unwrap());
    let lang_name: &'static str = lang.into();
    assert_eq!(lang, Lang::for_name(lang_name, &()).unwrap());
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Ada

Provides the ada language, supports these file types: ada.

Agda

Provides the agda language, supports these file types: agda.

Apex

Provides the apex language, supports these file types: apex.

Arduino

Provides the arduino language, supports these file types: arduino.

Asm

Provides the asm language, supports these file types: asm, vmasm.

Asm68k

Provides the asm68k language, supports these file types: m68k.

Astro

Provides the astro language, supports these file types: astro.

Authzed

Provides the authzed language, supports these file types: authzed.

Awk

Provides the awk language, supports these file types: awk.

Bash

Provides the bash language, supports these file types: sh, bash, ksh.

Bass

Provides the bass language, supports these file types: bass.

Beancount

Provides the beancount language, supports these file types: beancount.

Bib

Provides the bib language, supports these file types: bib.

Bibtex

Provides the bibtex language, supports these file types: bib.

Bicep

Provides the bicep language, supports these file types: bicep.

Bitbake

Provides the bitbake language, supports these file types: bitbake.

Blade

Provides the blade language, supports these file types: blade.

Blueprint

Provides the blueprint language, supports these file types: blueprint, blueprint.

Bp

Provides the bp language, supports these file types: bp.

Bpftrace

Provides the bpftrace language, supports these file types: bpftrace.

Brightscript

Provides the brightscript language, supports these file types: brightscript.

Bzl

Provides the bzl language, supports these file types: bzl.

C

Provides the c language, supports these file types: c.

C3

Provides the c3 language, supports these file types: c3.

CSharp

Provides the c_sharp language, supports these file types: csharp.

Caddy

Provides the caddy language, supports these file types: caddy.

Cairo

Provides the cairo language, supports these file types: cairo.

Capnp

Provides the capnp language, supports these file types: capnp.

Chatito

Provides the chatito language, supports these file types: chatito.

Circom

Provides the circom language, supports these file types: circom.

Clarity

Provides the clarity language, supports these file types: clarity, clarity.

Clojure

Provides the clojure language, supports these file types: clojure.

Cmake

Provides the cmake language, supports these file types: cmake.

Comment

Provides the comment language, supports no file types.

Cook

Provides the cook language, supports these file types: cook.

Corn

Provides the corn language, supports these file types: corn.

Cpon

Provides the cpon language, supports these file types: cpon.

Cpp

Provides the cpp language, supports these file types: cpp.

Csharp

Provides the csharp language, supports these file types: csharp.

Css

Provides the css language, supports these file types: css.

Csv

Provides the csv language, supports these file types: csv.

Cuda

Provides the cuda language, supports these file types: cuda.

Cue

Provides the cue language, supports these file types: cue.

Cylc

Provides the cylc language, supports these file types: cylc.

D

Provides the d language, supports these file types: d.

Dart

Provides the dart language, supports these file types: dart.

Desktop

Provides the desktop language, supports these file types: desktop.

Dhall

Provides the dhall language, supports these file types: dhall.

Diff

Provides the diff language, supports these file types: gitdiff.

Disassembly

Provides the disassembly language, supports these file types: disassembly.

Djot

Provides the djot language, supports these file types: djot.

Dockerfile

Provides the dockerfile language, supports these file types: dockerfile.

Dot

Provides the dot language, supports these file types: dot.

Doxygen

Provides the doxygen language, supports these file types: doxygen.

Dtd

Provides the dtd language, supports these file types: dtd.

Dts

Provides the dts language, supports these file types: dts.

Earthfile

Provides the earthfile language, supports these file types: earthfile.

Ebnf

Provides the ebnf language, supports these file types: ebnf.

Editorconfig

Provides the editorconfig language, supports these file types: editorconfig.

Eds

Provides the eds language, supports these file types: eds.

Eelixir

Provides the eelixir language, supports these file types: eelixir.

Ejs

Provides the ejs language, supports these file types: ejavascript.

Elisp

Provides the elisp language, supports these file types: elisp, elisp.

Elixir

Provides the elixir language, supports these file types: elixir.

Elm

Provides the elm language, supports these file types: elm.

Elsa

Provides the elsa language, supports these file types: elsa.

Elvish

Provides the elvish language, supports these file types: elvish.

Enforce

Provides the enforce language, supports these file types: enforce.

Erb

Provides the erb language, supports these file types: eruby.

Erlang

Provides the erlang language, supports these file types: erlang.

Eruby

Provides the eruby language, supports these file types: eruby.

Faust

Provides the faust language, supports these file types: faust.

Fennel

Provides the fennel language, supports these file types: fennel.

Fidl

Provides the fidl language, supports these file types: fidl.

Firrtl

Provides the firrtl language, supports these file types: firrtl.

Fish

Provides the fish language, supports these file types: fish.

Foam

Provides the foam language, supports these file types: foam.

Forth

Provides the forth language, supports these file types: forth.

Fortran

Provides the fortran language, supports these file types: fortran.

Fsd

Provides the fsd language, supports these file types: fsd.

Fsh

Provides the fsh language, supports these file types: fsh.

Fsharp

Provides the fsharp language, supports these file types: fsharp.

Func

Provides the func language, supports these file types: func.

Fusion

Provides the fusion language, supports these file types: fusion.

Gap

Provides the gap language, supports these file types: gap.

Gaptst

Provides the gaptst language, supports these file types: gaptst.

Gdresource

Provides the gdresource language, supports these file types: gdresource.

Gdscript

Provides the gdscript language, supports these file types: gdscript.

Gdshaderinc

Provides the gdshaderinc language, supports these file types: gdshaderinc.

Gitattributes

Provides the gitattributes language, supports these file types: gitattributes.

Gitcommit

Provides the gitcommit language, supports these file types: gitcommit.

Gitconfig

Provides the gitconfig language, supports these file types: gitconfig.

Gitdiff

Provides the gitdiff language, supports these file types: gitdiff.

Gitignore

Provides the gitignore language, supports these file types: gitignore.

Gitrebase

Provides the gitrebase language, supports these file types: gitrebase.

Gleam

Provides the gleam language, supports these file types: gleam.

GlimmerJavascript

Provides the glimmer_javascript language, supports these file types: javascript.glimmer.

GlimmerTypescript

Provides the glimmer_typescript language, supports these file types: typescript.glimmer.

Glsl

Provides the glsl language, supports these file types: glsl.

Gn

Provides the gn language, supports these file types: gn.

Gnuplot

Provides the gnuplot language, supports these file types: gnuplot.

Go

Provides the go language, supports these file types: go.

Goctl

Provides the goctl language, supports these file types: goctl.

Gomod

Provides the gomod language, supports these file types: gomod.

Gosum

Provides the gosum language, supports these file types: gosum.

Gotmpl

Provides the gotmpl language, supports these file types: gotmpl.

Gowork

Provides the gowork language, supports these file types: gowork.

Gpg

Provides the gpg language, supports these file types: gpg.

Graphql

Provides the graphql language, supports these file types: graphql.

Gren

Provides the gren language, supports these file types: gren.

Groovy

Provides the groovy language, supports these file types: groovy.

Groq

Provides the groq language, supports these file types: groq.

Gstlaunch

Provides the gstlaunch language, supports these file types: gstlaunch.

Hack

Provides the hack language, supports these file types: hack.

Handlebars

Provides the handlebars language, supports these file types: handlebars.

Hare

Provides the hare language, supports these file types: hare.

Haskell

Provides the haskell language, supports these file types: haskell.

Haskellpersistent

Provides the haskellpersistent language, supports these file types: haskellpersistent.

Haxe

Provides the haxe language, supports these file types: haxe, haxe.

Hcl

Provides the hcl language, supports these file types: hcl.

Heex

Provides the heex language, supports these file types: heex.

Helm

Provides the helm language, supports these file types: helm.

Help

Provides the help language, supports these file types: help.

Hexdump

Provides the hexdump language, supports these file types: hexdump.

Hjson

Provides the hjson language, supports these file types: hjson.

Hlsl

Provides the hlsl language, supports these file types: hlsl.

Hlsplaylist

Provides the hlsplaylist language, supports these file types: hlsplaylist.

Hocon

Provides the hocon language, supports these file types: hocon.

Hoon

Provides the hoon language, supports these file types: hoon.

Html

Provides the html language, supports these file types: html.

Htmlangular

Provides the htmlangular language, supports these file types: htmlangular.

Htmldjango

Provides the htmldjango language, supports these file types: htmldjango.

Http

Provides the http language, supports these file types: http.

Hurl

Provides the hurl language, supports these file types: hurl.

Hyprlang

Provides the hyprlang language, supports these file types: hyprlang.

Idl

Provides the idl language, supports these file types: idl.

Idris2

Provides the idris2 language, supports these file types: idris2.

Ini

Provides the ini language, supports these file types: ini, systemd.

Inko

Provides the inko language, supports these file types: inko.

Ispc

Provides the ispc language, supports these file types: ispc.

Janet

Provides the janet language, supports these file types: janet.

Java

Provides the java language, supports these file types: java.

Javadoc

Provides the javadoc language, supports these file types: javadoc.

Javascript

Provides the javascript language, supports these file types: javascript, jsx.

Jinja

Provides the jinja language, supports these file types: jinja.

JinjaInline

Provides the jinja_inline language, supports these file types: jinja_inline.

Jproperties

Provides the jproperties language, supports these file types: jproperties.

Jq

Provides the jq language, supports these file types: jq.

Jsdoc

Provides the jsdoc language, supports no file types.

Json

Provides the json language, supports these file types: json.

Json5

Provides the json5 language, supports these file types: json5.

Jsonc

Provides the jsonc language, supports these file types: jsonc.

Jsonnet

Provides the jsonnet language, supports these file types: jsonnet.

Julia

Provides the julia language, supports these file types: julia.

Just

Provides the just language, supports these file types: just.

Kcl

Provides the kcl language, supports these file types: kcl.

Kconfig

Provides the kconfig language, supports these file types: kconfig.

Kdl

Provides the kdl language, supports these file types: kdl.

Kitty

Provides the kitty language, supports these file types: kitty.

Kos

Provides the kos language, supports these file types: kos.

Kotlin

Provides the kotlin language, supports these file types: kotlin.

Koto

Provides the koto language, supports these file types: koto.

Kusto

Provides the kusto language, supports these file types: kusto.

Lalrpop

Provides the lalrpop language, supports no file types.

Latex

Provides the latex language, supports these file types: tex.

Ld

Provides the ld language, supports these file types: ld.

Ledger

Provides the ledger language, supports these file types: ledger.

Leo

Provides the leo language, supports these file types: leo.

Liquid

Provides the liquid language, supports these file types: liquid.

Liquidsoap

Provides the liquidsoap language, supports these file types: liquidsoap.

Lisp

Provides the lisp language, supports these file types: lisp.

Llvm

Provides the llvm language, supports these file types: llvm.

Lua

Provides the lua language, supports these file types: lua.

Luadoc

Provides the luadoc language, supports these file types: luadoc.

Luap

Provides the luap language, supports no file types.

Luau

Provides the luau language, supports these file types: luau.

Magik

Provides the magik language, supports these file types: magik.

Make

Provides the make language, supports these file types: make, automake.

Markdown

Provides the markdown language, supports these file types: markdown.

MarkdownInline

Provides the markdown_inline language, supports no file types.

Matlab

Provides the matlab language, supports these file types: matlab.

Menhir

Provides the menhir language, supports these file types: menhir.

Mermaid

Provides the mermaid language, supports these file types: mermaid.

Meson

Provides the meson language, supports these file types: meson.

Mlir

Provides the mlir language, supports these file types: mlir.

Muttrc

Provides the muttrc language, supports these file types: muttrc.

Nasm

Provides the nasm language, supports these file types: nasm.

Nginx

Provides the nginx language, supports these file types: nginx.

Nickel

Provides the nickel language, supports these file types: nickel.

Nim

Provides the nim language, supports these file types: nim.

NimFormatString

Provides the nim_format_string language, supports these file types: nim_format_string.

Ninja

Provides the ninja language, supports these file types: ninja.

Nix

Provides the nix language, supports these file types: nix.

Nqc

Provides the nqc language, supports these file types: nqc.

Nu

Provides the nu language, supports these file types: nu.

Objc

Provides the objc language, supports these file types: objc.

Objdump

Provides the objdump language, supports these file types: objdump.

Ocaml

Provides the ocaml language, supports these file types: ocaml.

OcamlInterface

Provides the ocaml_interface language, supports these file types: ocamlinterface.

Ocamlinterface

Provides the ocamlinterface language, supports these file types: ocamlinterface.

Ocamllex

Provides the ocamllex language, supports these file types: ocamllex.

Odin

Provides the odin language, supports these file types: odin.

Pascal

Provides the pascal language, supports these file types: pascal.

Passwd

Provides the passwd language, supports these file types: passwd.

Pbtxt

Provides the pbtxt language, supports these file types: pbtxt.

Pem

Provides the pem language, supports these file types: pem.

Perl

Provides the perl language, supports these file types: perl.

Pgn

Provides the pgn language, supports these file types: pgn.

Php

Provides the php language, supports these file types: php.

PhpOnly

Provides the php_only language, supports no file types.

Phpdoc

Provides the phpdoc language, supports these file types: phpdoc.

Pioasm

Provides the pioasm language, supports these file types: pioasm.

Pkl

Provides the pkl language, supports these file types: pkl.

Po

Provides the po language, supports these file types: po.

Pod

Provides the pod language, supports these file types: pod.

Poefilter

Provides the poefilter language, supports these file types: poefilter.

Pony

Provides the pony language, supports these file types: pony.

Printf

Provides the printf language, supports no file types.

Prisma

Provides the prisma language, supports these file types: prisma.

Problog

Provides the problog language, supports these file types: problog.

Prolog

Provides the prolog language, supports these file types: prolog.

Promql

Provides the promql language, supports these file types: promql.

Proto

Provides the proto language, supports these file types: proto.

Prql

Provides the prql language, supports these file types: prql.

Ps1

Provides the ps1 language, supports these file types: ps1.

Psv

Provides the psv language, supports these file types: psv.

Pug

Provides the pug language, supports these file types: pug.

Puppet

Provides the puppet language, supports these file types: puppet.

Purescript

Provides the purescript language, supports these file types: purescript.

Pymanifest

Provides the pymanifest language, supports these file types: pymanifest.

Python

Provides the python language, supports these file types: python.

Ql

Provides the ql language, supports these file types: ql.

Qml

Provides the qml language, supports these file types: qml.

Qmldir

Provides the qmldir language, supports these file types: qmldir.

Query

Provides the query language, supports these file types: query.

R

Provides the r language, supports these file types: r.

Racket

Provides the racket language, supports these file types: racket.

Ralph

Provides the ralph language, supports these file types: ralph.

Rasi

Provides the rasi language, supports these file types: rasi.

Razor

Provides the razor language, supports these file types: razor.

Rbs

Provides the rbs language, supports these file types: rbs.

Re2c

Provides the re2c language, supports these file types: re2c.

Readline

Provides the readline language, supports these file types: readline.

Regex

Provides the regex language, supports no file types.

Rego

Provides the rego language, supports these file types: rego.

Requirements

Provides the requirements language, supports these file types: requirements.

Rescript

Provides the rescript language, supports these file types: rescript.

Rifleconf

Provides the rifleconf language, supports these file types: rifleconf.

Rnoweb

Provides the rnoweb language, supports these file types: rnoweb.

Robot

Provides the robot language, supports these file types: robot.

Robots

Provides the robots language, supports these file types: robots.

Roc

Provides the roc language, supports these file types: roc.

Ron

Provides the ron language, supports these file types: ron.

Rst

Provides the rst language, supports these file types: rst.

Ruby

Provides the ruby language, supports these file types: ruby.

Runescript

Provides the runescript language, supports these file types: runescript.

Rush

Provides the rush language, supports these file types: rush.

Rust

Provides the rust language, supports these file types: rust.

Scala

Provides the scala language, supports these file types: scala, sbt.

Scfg

Provides the scfg language, supports these file types: scfg.

Scheme

Provides the scheme language, supports these file types: scheme.

Scss

Provides the scss language, supports these file types: scss.

Sface

Provides the sface language, supports these file types: surface.

Sflog

Provides the sflog language, supports these file types: sflog.

Shaderslang

Provides the shaderslang language, supports these file types: shaderslang.

Slim

Provides the slim language, supports these file types: slim.

Slint

Provides the slint language, supports these file types: slint.

Smali

Provides the smali language, supports these file types: smali.

Smithy

Provides the smithy language, supports these file types: smithy.

Snakemake

Provides the snakemake language, supports these file types: snakemake.

Solidity

Provides the solidity language, supports these file types: solidity.

Soql

Provides the soql language, supports these file types: soql.

Sosl

Provides the sosl language, supports these file types: sosl.

Sourcepawn

Provides the sourcepawn language, supports these file types: sourcepawn.

Sparql

Provides the sparql language, supports these file types: sparql.

Sproto

Provides the sproto language, supports these file types: sproto.

Sql

Provides the sql language, supports these file types: sql, mysql.

Squirrel

Provides the squirrel language, supports these file types: squirrel.

Sshconfig

Provides the sshconfig language, supports these file types: sshconfig.

Strace

Provides the strace language, supports these file types: strace.

Styled

Provides the styled language, supports these file types: styled.

Supercollider

Provides the supercollider language, supports these file types: supercollider.

Superhtml

Provides the superhtml language, supports these file types: superhtml.

Svelte

Provides the svelte language, supports these file types: svelte.

Sway

Provides the sway language, supports these file types: sway.

Swift

Provides the swift language, supports these file types: swift.

Sxhkdrc

Provides the sxhkdrc language, supports these file types: sxhkdrc.

Systemtap

Provides the systemtap language, supports these file types: systemtap.

Tablegen

Provides the tablegen language, supports these file types: tablegen.

Tact

Provides the tact language, supports these file types: tact.

Tal

Provides the tal language, supports these file types: tal.

Tape

Provides the tape language, supports these file types: vhs.

Tcl

Provides the tcl language, supports these file types: tcl.

Teal

Provides the teal language, supports these file types: teal.

Templ

Provides the templ language, supports these file types: templ.

Tera

Provides the tera language, supports these file types: tera.

Terraform

Provides the terraform language, supports these file types: terraform.

Tex

Provides the tex language, supports these file types: tex.

Thrift

Provides the thrift language, supports these file types: thrift.

Tiger

Provides the tiger language, supports these file types: tiger.

Tla

Provides the tla language, supports these file types: tla.

Tmux

Provides the tmux language, supports these file types: tmux.

Todotxt

Provides the todotxt language, supports these file types: todotxt.

Toml

Provides the toml language, supports these file types: toml.

Trace32

Provides the trace32 language, supports these file types: trace32.

Tsv

Provides the tsv language, supports these file types: tsv.

Tsx

Provides the tsx language, supports these file types: tsx.

Turtle

Provides the turtle language, supports these file types: turtle.

Twig

Provides the twig language, supports these file types: twig.

Typescript

Provides the typescript language, supports these file types: typescript.

Typespec

Provides the typespec language, supports these file types: typespec.

Typoscript

Provides the typoscript language, supports these file types: typoscript.

Typst

Provides the typst language, supports these file types: typst.

Udevrules

Provides the udevrules language, supports these file types: udevrules.

Ungrammar

Provides the ungrammar language, supports these file types: ungrammar.

Unison

Provides the unison language, supports these file types: unison.

Ursa

Provides the ursa language, supports these file types: ursa.

Usd

Provides the usd language, supports these file types: usd.

V

Provides the v language, supports these file types: v.

Vala

Provides the vala language, supports these file types: vala.

Verilog

Provides the verilog language, supports these file types: systemverilog, systemverilog.

Vhdl

Provides the vhdl language, supports these file types: vhdl.

Vim

Provides the vim language, supports these file types: vim.

Vrl

Provides the vrl language, supports these file types: vrl.

Vto

Provides the vto language, supports these file types: vto.

Vue

Provides the vue language, supports these file types: vue.

Wat

Provides the wat language, supports these file types: wat.

Wgsl

Provides the wgsl language, supports these file types: wgsl.

WgslBevy

Provides the wgsl_bevy language, supports these file types: wgsl_bevy.

Wing

Provides the wing language, supports these file types: wing.

Wit

Provides the wit language, supports these file types: wit.

Wxml

Provides the wxml language, supports these file types: wxml.

Xcompose

Provides the xcompose language, supports these file types: xcompose.

Xdefaults

Provides the xdefaults language, supports these file types: xdefaults.

Xml

Provides the xml language, supports these file types: xml.

Yaml

Provides the yaml language, supports these file types: yaml.

Yang

Provides the yang language, supports these file types: yang.

Yuck

Provides the yuck language, supports these file types: yuck.

Zathurarc

Provides the zathurarc language, supports these file types: zathurarc.

Zig

Provides the zig language, supports these file types: zig.

Ziggy

Provides the ziggy language, supports these file types: ziggy.

ZiggySchema

Provides the ziggy_schema language, supports these file types: ziggy_schema.

Zsh

Provides the zsh language, supports these file types: zsh.

Implementations

impl Lang

pub fn get(&self) -> Language

Get an instance of the respective Language.

pub fn get_config(&self) -> Result<HighlightConfiguration>

Create an instance of the corresponding HighlightConfiguration.

pub fn highlights_query(&self) -> &'static str

Get the highlights query for this language.

pub fn injections_query(&self) -> &'static str

Get the injections query for this language.

pub fn locals_query(&self) -> &'static str

Get the locals query for this language.